fix(frontend): インストールしたテーマがテーマ一覧にすぐ反映されない
